Implementada a função girar na simulação. obs: o robo da simulação irá girar apenas quando nenhum robo real estiver conectado.

+++ src/robotinterface/robot/device/Compass.java
@@ -59,7 +59,7 @@ public class Compass extends Device implements VirtualDevice {
         return 3;
     }
 
-    public double getAlpha() {
+    public int getAlpha() {
         return alpha;
     }
